Move the site updates reducer to the sites.items reducer
The only action that needs any treatment is SITE_PLUGIN_UPDATED
