--- client/state/sites/updates/schema.js
@@ -1,19 +1,0 @@
-/** @format */
-export const itemsSchema = {
-	type: 'object',
-	additionalProperties: false,
-	patternProperties: {
-		'^\\d+$': {
-			type: 'object',
-			properties: {
-				jp_version: { type: 'string' },
-				plugins: { type: 'number' },
-				themes: { type: 'number' },
-				total: { type: 'number' },
-				translations: { type: 'number' },
-				wordpress: { type: 'number' },
-				wp_version: { type: 'string' },
-			},
-		},
-	},
-};
